+++ app/scripts/main/components/webview.js
@@ -6,10 +6,8 @@
  * Node
  * @constant
  */
-const fs = require('fs');
 const path = require('path');
 const url = require('url');
-const util = require('util');
 
 /**
  * Modules
@@ -44,7 +42,7 @@ const logger = require(path.join(appRootPath, 'lib', 'logger'))({ write: true })
 const body = document.querySelector('body');
 const webview = document.getElementById('webview');
 const spinner = document.getElementById('spinner');
-const status = document.getElementById('spinner__text');
+const statustext = document.getElementById('spinner__text');
 const controlsExtra = document.getElementById('controls-extra');
 const buttons = {
     home: {
@@ -144,10 +142,21 @@ webview.addEventListener('did-navigate-in-page', (ev) => {
  * @listens webview#new-window
  */
 webview.addEventListener('new-window', (ev) => {
-    logger.debug('webview#new-window', 'ev.url:', ev.url);
+    logger.debug('webview#new-window');
 
     ev.preventDefault();
-    webview.loadURL(ev.url);
+
+    let domain = parseDomain(ev.url)['domain'] || '';
+
+    if (domain === 'pushbullet') {
+        // Internal Link
+        logger.info('webview#new-window', 'opening internal url:', ev.url);
+        webview.loadURL(ev.url);
+    } else {
+        // External Link
+        logger.info('webview#new-window', 'opening external url:', ev.url);
+        remote.shell.openExternal(ev.url);
+    }
 });
 
 /**
@@ -195,12 +204,12 @@ webview.addEventListener('load-commit', (ev) => {
 });
 
 /**
+ * CSS Injection
  * @listens webview#load-commit
  */
-webview.addEventListener('load-commit', (ev) => {
+webview.addEventListener('load-commit', () => {
     logger.debug('webview#load-commit');
 
-    // CSS Injection
     domHelper.injectCSS(webview, path.join(appRootPath, 'app', 'styles', 'pushbullet.css'));
 });
 
@@ -242,23 +251,25 @@ webview.addEventListener('ipc-message', (ev) => {
         case 'account':
             switch (message) {
                 case 'login':
-                logger.info('account', 'login');
-                domHelper.setText(status, 'logged in');
+                    logger.info('account', 'login');
+                    domHelper.setText(statustext, 'logged in');
+                    break;
             }
+            break;
         case 'network':
             const didDisconnect = ev.args[1];
             switch (message) {
                 case 'offline':
                     logger.info('network', 'offline');
                     presentSpinner();
-                    domHelper.setText(status, 'connecting...');
+                    domHelper.setText(statustext, 'connecting...');
                     break;
                 case 'online':
                     logger.info('network', 'online');
-                    domHelper.setText(status, 'connected');
+                    domHelper.setText(statustext, 'connected');
                     if (Boolean(didDisconnect)) {
                         logger.info('network', 'reconnecting...');
-                        domHelper.setText(status, 'reconnecting');
+                        domHelper.setText(statustext, 'reconnecting');
                         webview.reloadIgnoringCache();
                     }
                     dismissSpinner();